EDUCATION TUBE
This is good for your Kids around 2 to 12 Years Old. And it Is the same as youtube but this is different. Our Platform Is Very Powerful No-one can See or diwnload without the QR Code . Your child is allow To Upload Their Videos And All the other Children In This World Will Watch your son/dauther's video I Hope you Enjoy the App.